Revert "sample_dev"
