+++ src/main/kotlin/com/chigirh/tools/common/file/FileWrapper.kt
@@ -1,10 +1,8 @@
 package com.chigirh.tools.common.file
 
-import com.chigirh.tools.common.file.FileUtil.write
-import java.nio.charset.Charset
 import java.nio.file.Path
+import kotlin.io.path.bufferedWriter
 import kotlin.io.path.readLines
-import kotlin.io.path.readText
 
 /**
  * Simple file wrapper.
@@ -12,13 +10,20 @@ import kotlin.io.path.readText
  */
 class FileWrapper(
     private val file: Path,
-    private val charset: Charset = Charsets.UTF_8
 ) {
-    fun write(text: String) = file.write(text, charset)
+    fun write(text: String) {
+        file.bufferedWriter().use { bw ->
+            bw.write(text)
+            true
+        }
+    }
 
-    fun write(texts: List<String>) = file.write(texts, charset)
+    fun write(texts: List<String>) {
+        file.bufferedWriter().use { bw ->
+            texts.forEach { bw.write(it) }
+            true
+        }
+    }
 
-    fun readLines() = file.readLines(charset)
-
-    fun readText() = file.readText(charset)
+    fun readLines() = file.readLines()
 }
