+++ src/main/kotlin/com/chigirh/tools/common/pipeline/CommandPipeLineBase.kt
@@ -6,14 +6,12 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.launch
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.time.LocalDateTime
-import java.time.temporal.ChronoUnit
 
 abstract class CommandPipeLineBase<CMD : PipeLineCommand>(
-    protected val pipeLineName: String,
-    protected val pipeLineSize: Int,
-    protected val commandNames: List<String>,
-    protected val option: CommandPipeLineOption = CommandPipeLineOption.DEFAULT,
+    private val isParallel: Boolean = true,
+    private val pipeLineName: String,
+    private val pipeLineSize: Int,
+    private val commandNames: List<String>,
     private val channels: MutableMap<String, Channel<PipeLineMessenger<CMD>>> = HashMap()
 ) : CommandPipeline {
     private var isActive: Boolean = false
@@ -25,19 +23,16 @@ abstract class CommandPipeLineBase<CMD : PipeLineCommand>(
 
         isActive = true
         for (command in commandNames) {
-            val channel = Channel<PipeLineMessenger<CMD>>()
             repeat(pipeLineSize) {
+                val channel = Channel<PipeLineMessenger<CMD>>()
                 channels[command] = channel
-                val task = createSlave(channel, it)
+                val task = consumer(channel)
                 task.run()
             }
         }
     }
 
-    abstract fun createSlave(
-        channel: Channel<PipeLineMessenger<CMD>>,
-        slaveNumber: Int,
-    ): PipeLineSlave<CMD>
+    abstract fun consumer(channel: Channel<PipeLineMessenger<CMD>>): PipeLineConsumer<CMD>
 
     /**
      * Flowing for pipe line.
@@ -47,29 +42,19 @@ abstract class CommandPipeLineBase<CMD : PipeLineCommand>(
      */
     fun flowing(task: PipeLineTask<CMD>) =
         CoroutineScope(Dispatchers.Default).launch {
-            if (option.isParallel) parallel(task) else series(task)
+            if (isParallel) parallel(task) else series(task)
         }
 
     private suspend fun parallel(task: PipeLineTask<CMD>) {
         val noticeChannel = Channel<Int>()
-
-        val startedAt = LocalDateTime.now()
-
         task.commands.forEach {
             CoroutineScope(Dispatchers.Default).launch {
-                val messenger = PipeLineMessenger(task.taskName, noticeChannel, it)
+                val messenger = PipeLineMessenger(noticeChannel, it)
                 channels[it.commandName]!!.send(messenger)
             }
         }
         repeat(task.commands.size) {
             noticeChannel.receive()
-        }
-
-        val endedAt = LocalDateTime.now()
-
-        val processingTime = ChronoUnit.MILLIS.between(startedAt, endedAt)
-        if (option.outputLog && option.exceedLimit < processingTime) {
-            logger.info("$pipeLineName -- message:exceed log ${task.taskName}->${processingTime}ms")
         }
     }
 
@@ -77,7 +62,7 @@ abstract class CommandPipeLineBase<CMD : PipeLineCommand>(
         val noticeChannel = Channel<Int>()
         task.commands.sortedBy { it.sequence }.forEach {
             CoroutineScope(Dispatchers.Default).launch {
-                val messenger = PipeLineMessenger(task.taskName, noticeChannel, it)
+                val messenger = PipeLineMessenger(noticeChannel, it)
                 channels[it.commandName]!!.send(messenger)
             }
             noticeChannel.receive()
@@ -88,3 +73,4 @@ abstract class CommandPipeLineBase<CMD : PipeLineCommand>(
         val logger: Logger = LoggerFactory.getLogger(CommandPipeLineBase.javaClass)
     }
 }
+
